#315 rename fgs_communicator
