+++ src/artemis/external_interaction/communicator_callbacks.py
